[Quantization] Implement TP weight switching support and refactor related methods
