+++ vllm_ascend/quantization/methods/base.py
@@ -18,57 +18,32 @@
 
 from abc import ABC, abstractmethod
 from collections.abc import Callable
-from dataclasses import dataclass, field
-from typing import Any, ClassVar
+from typing import Any
 
 import torch
 
 from vllm_ascend.quantization.quant_type import QuantType
+from vllm_ascend.quantization.tp_weight_switch import (
+    TPWeightGatherPart,
+    TPWeightGatherSpec,
+    TPWeightRepeatPart,
+    TPWeightRepeatSpec,
+    TPWeightSwitchMixin,
+    TPWeightSwitchState,
+)
 
-
-@dataclass(frozen=True)
-class TPWeightGatherSpec:
-    """Tensor attribute that should be all-gathered across TP ranks."""
-
-    attr_name: str
-    gather_dim: int = 0
-
-
-@dataclass(frozen=True)
-class TPWeightRepeatSpec:
-    """Tensor attribute that should be expanded by local repeat."""
-
-    attr_name: str
-    repeat_dim: int = 0
-
-
-@dataclass
-class TPWeightGatherPart:
-    """Runtime tensors for one all-gathered TP attribute."""
-
-    spec: TPWeightGatherSpec
-    tp_tensor: torch.Tensor
-    gather_input: torch.Tensor
-    gather_output: torch.Tensor
-    full_tensor: torch.Tensor
-
-
-@dataclass
-class TPWeightRepeatPart:
-    """Runtime tensors for one repeated TP attribute."""
-
-    spec: TPWeightRepeatSpec
-    tp_tensor: torch.Tensor
-    full_tensor: torch.Tensor
-
-
-@dataclass
-class TPWeightSwitchState:
-    """Reusable buffers for switching a layer between TP and full weights."""
-
-    gather_parts: dict[str, TPWeightGatherPart] = field(default_factory=dict)
-    repeat_parts: dict[str, TPWeightRepeatPart] = field(default_factory=dict)
-    handles: list[torch.distributed.Work] = field(default_factory=list)
+__all__ = [
+    "AscendAttentionScheme",
+    "AscendLinearScheme",
+    "AscendMoEScheme",
+    "QuantType",
+    "TPWeightGatherPart",
+    "TPWeightGatherSpec",
+    "TPWeightRepeatPart",
+    "TPWeightRepeatSpec",
+    "TPWeightSwitchMixin",
+    "TPWeightSwitchState",
+]
 
 
 def get_moe_num_logical_experts(
@@ -85,16 +60,13 @@ def get_moe_num_logical_experts(
     return int(num_experts - global_redundant_expert_num - num_shared_experts)
 
 
-class AscendLinearScheme(ABC):
+class AscendLinearScheme(TPWeightSwitchMixin, ABC):
     """Base class for all linear quantization schemes.
 
     Subclasses must implement get_weight() and apply() methods.
     Other methods have default implementations that return empty dicts
     or do nothing.
     """
-
-    tp_weight_gather_specs: ClassVar[tuple[TPWeightGatherSpec, ...]] = ()
-    tp_weight_repeat_specs: ClassVar[tuple[TPWeightRepeatSpec, ...]] = ()
 
     @abstractmethod
     def get_weight(self, input_size: int, output_size: int, params_dtype: torch.dtype) -> dict[str, Any]:
@@ -175,169 +147,6 @@ class AscendLinearScheme(ABC):
             layer: The linear layer module.
         """
         return
-
-    @staticmethod
-    def split_tensor_for_tp(
-        tensor: torch.Tensor,
-        tp_size: int,
-        tp_rank: int,
-        dim: int = 0,
-        contiguous: bool = True,
-    ) -> torch.Tensor:
-        """Slice one TP shard from a full tensor."""
-        dim = dim if dim >= 0 else tensor.dim() + dim
-        if tensor.shape[dim] % tp_size != 0:
-            raise RuntimeError(
-                "Cannot split tensor for TP because the target dimension is not divisible by TP size: "
-                f"shape={tuple(tensor.shape)}, dim={dim}, tp_size={tp_size}."
-            )
-        shard_size = tensor.shape[dim] // tp_size
-        shard = tensor.narrow(dim, tp_rank * shard_size, shard_size)
-        return shard.contiguous() if contiguous else shard
-
-    def enable_tp_weight_switch(
-        self,
-        layer: torch.nn.Module,
-        tp_size: int,
-        *,
-        pool: dict[Any, torch.Tensor] | None = None,
-        pool_key_prefix: Any | None = None,
-        clone_tp_tensors: bool = False,
-    ) -> TPWeightSwitchState:
-        """Enable TP/full weight switching for one linear layer.
-
-        This is deliberately an explicit feature entry point. It allocates the
-        full-weight and repeated-parameter buffers only when a caller needs to
-        switch a layer out of its normal TP layout.
-        """
-        state = TPWeightSwitchState()
-        gather_specs = self.tp_weight_gather_specs
-        repeat_specs = self.tp_weight_repeat_specs
-
-        for spec in gather_specs:
-            tensor = getattr(layer, spec.attr_name, None)
-            if tensor is None:
-                raise RuntimeError(
-                    f"{type(self).__name__} declares TP gather attr {spec.attr_name!r}, "
-                    f"but layer {getattr(layer, 'prefix', layer)} does not have it."
-                )
-            dim = spec.gather_dim if spec.gather_dim >= 0 else tensor.dim() + spec.gather_dim
-            if dim < 0 or dim >= tensor.dim():
-                raise RuntimeError(
-                    f"Invalid TP gather dim {spec.gather_dim} for attr {spec.attr_name!r} "
-                    f"with shape {tuple(tensor.shape)}."
-            )
-
-            tp_tensor = tensor.clone().detach().contiguous() if clone_tp_tensors else tensor.detach()
-            if clone_tp_tensors:
-                with torch.no_grad():
-                    tensor.set_(tp_tensor)
-            full_shape_list = list(tp_tensor.shape)
-            full_shape_list[dim] *= tp_size
-            full_shape = tuple(full_shape_list)
-            if dim == 0:
-                gather_input = tp_tensor
-            else:
-                gather_input = torch.movedim(tp_tensor, dim, 0).contiguous()
-            gather_shape = (full_shape[dim], *full_shape[:dim], *full_shape[dim + 1 :])
-            pool_key = (
-                pool_key_prefix,
-                spec.attr_name,
-                tp_tensor.device.type,
-                tp_tensor.device.index,
-                tp_tensor.dtype,
-                dim,
-                full_shape,
-            )
-            if pool is None:
-                gather_output = torch.empty(gather_shape, dtype=tp_tensor.dtype, device=tp_tensor.device)
-            else:
-                gather_output = pool.get(pool_key)
-                if gather_output is None:
-                    gather_output = torch.empty(gather_shape, dtype=tp_tensor.dtype, device=tp_tensor.device)
-                    pool[pool_key] = gather_output
-            if dim == 0:
-                full_tensor = gather_output
-            else:
-                full_tensor = torch.movedim(gather_output, 0, dim)
-            state.gather_parts[spec.attr_name] = TPWeightGatherPart(
-                spec=spec,
-                tp_tensor=tp_tensor,
-                gather_input=gather_input,
-                gather_output=gather_output,
-                full_tensor=full_tensor,
-            )
-
-        for spec in repeat_specs:
-            tensor = getattr(layer, spec.attr_name, None)
-            if tensor is None:
-                raise RuntimeError(
-                    f"{type(self).__name__} declares TP repeat attr {spec.attr_name!r}, "
-                    f"but layer {getattr(layer, 'prefix', layer)} does not have it."
-                )
-            dim = spec.repeat_dim if spec.repeat_dim >= 0 else tensor.dim() + spec.repeat_dim
-            if dim < 0 or dim >= tensor.dim():
-                raise RuntimeError(
-                    f"Invalid TP repeat dim {spec.repeat_dim} for attr {spec.attr_name!r} "
-                    f"with shape {tuple(tensor.shape)}."
-                )
-            repeats = [1] * tensor.dim()
-            repeats[dim] = tp_size
-            tp_tensor = tensor.detach()
-            state.repeat_parts[spec.attr_name] = TPWeightRepeatPart(
-                spec=spec,
-                tp_tensor=tp_tensor,
-                full_tensor=tp_tensor.repeat(*repeats),
-            )
-
-        return state
-
-    def all_gather_tp_weight(
-        self,
-        state: TPWeightSwitchState,
-        group: Any,
-        *,
-        async_op: bool = True,
-    ) -> None:
-        """All-gather every TP-sharded tensor in ``state``."""
-        from vllm_ascend.distributed.utils import all_gather_async
-
-        if state.handles:
-            raise RuntimeError("TP weight all-gather is still pending; wait before launching another one.")
-        for part in state.gather_parts.values():
-            _, handle = all_gather_async(
-                part.gather_input,
-                group,
-                output=part.gather_output,
-                async_op=async_op,
-            )
-            if handle is not None:
-                state.handles.append(handle)
-
-    @staticmethod
-    def wait_tp_weight_all_gather(state: TPWeightSwitchState) -> None:
-        try:
-            for handle in state.handles:
-                handle.wait()
-        finally:
-            state.handles.clear()
-
-    def switch_tp_weight(
-        self,
-        layer: torch.nn.Module,
-        state: TPWeightSwitchState,
-        *,
-        use_full_weight: bool,
-    ) -> None:
-        """Switch layer tensor attributes between TP-local and full views."""
-        for attr_name, part in state.gather_parts.items():
-            target = part.full_tensor if use_full_weight else part.tp_tensor
-            with torch.no_grad():
-                getattr(layer, attr_name).set_(target)
-        for attr_name, part in state.repeat_parts.items():
-            target = part.full_tensor if use_full_weight else part.tp_tensor
-            with torch.no_grad():
-                getattr(layer, attr_name).set_(target)
 
 
 class AscendAttentionScheme(ABC):
